+++ hemi/rpc_grpc_suite.go
@@ -168,7 +168,7 @@ func (c *grpcConn) IsUDS() bool { return c.gate.IsUDS() }
 func (c *grpcConn) IsTLS() bool { return c.gate.IsTLS() }
 
 func (c *grpcConn) MakeTempName(to []byte, unixTime int64) int {
-	return makeTempName(to, int64(c.gate.server.Stage().ID()), c.id, unixTime, c.counter.Add(1))
+	return makeTempName(to, c.gate.server.Stage().ID(), c.id, unixTime, c.counter.Add(1))
 }
 
 //func (c *grpcConn) rpcServer() *grpcServer { return c.server }
