+++ hemi/rpc_hrpc_suite.go
@@ -171,7 +171,7 @@ func (c *hrpcConn) IsUDS() bool { return c.gate.IsUDS() }
 func (c *hrpcConn) IsTLS() bool { return c.gate.IsTLS() }
 
 func (c *hrpcConn) MakeTempName(to []byte, unixTime int64) int {
-	return makeTempName(to, int64(c.gate.server.Stage().ID()), c.id, unixTime, c.counter.Add(1))
+	return makeTempName(to, c.gate.server.Stage().ID(), c.id, unixTime, c.counter.Add(1))
 }
 
 //func (c *hrpcConn) rpcServer() *hrpcServer { return c.server }
